mydeluxetable.sty: don't override ucthesis caption commands
Things start getting ugly here. The deluxetable style overrides
various caption commands that are also defined in
(my)ucthesis.cls. These cause weird things to happen for figure
numbering and things outside of the deluxetable package. If we turn
off the caption setup in (my)deluxetable, everything still seems to
work OK, so let's go ahead and do that.

The customizations in (my)deluxetable seem to mostly have to do with
splitting captions across pages, which I seriously doubt the
(my)ucthesis commands can do. So if you have that situation, this
commit might break all that. If that happens, you're the lucky person
that gets to figure out how to make everybody play together nicely.