thesis.tex: suppress hypertex warnings about duplicate page anchors
This apparently comes from the title and abstract pages, which aren't
going to be linked to, so we can temporarily turn off the creation of
page anchors for those pages.

Also remove unneeded page tweaking commands in the front matter. The
correct magic is already done in the document class, as you might
expect.