pwucthesis.cls: unbold title on copyright page too
Making this consistent with the other pages that involve the title.
There's no reference copyright page on the dissertation guidelines so
I don't know if there are any other tweaks to make at the moment. It
looks like all of these fixed-format pages are a bit more flexible
than the let on so it's probably fine.